requirements.txt
streamlit
openai>=1.0.0
python-pptx
python-dotenv
